+++ community/consistency-check/src/test/java/org/neo4j/consistency/checking/cache/PackedMultiFieldCacheTest.java
@@ -33,7 +33,7 @@ class PackedMultiFieldCacheTest
     void shouldPutValuesIntoSlots()
     {
         // GIVEN
-        PackedMultiFieldCache cache = new PackedMultiFieldCache( defaultByteArray( (long) 20, INSTANCE ), 8, 16, 24, 32, 1 );
+        PackedMultiFieldCache cache = new PackedMultiFieldCache( defaultByteArray( 20, INSTANCE ), 8, 16, 24, 32, 1 );
         int index = 10;
         long[] values = new long[] {3, 100, 12345, 67890, 0};
 
@@ -51,7 +51,7 @@ class PackedMultiFieldCacheTest
     void shouldHaveCorrectDefaultValues()
     {
         // GIVEN
-        PackedMultiFieldCache cache = new PackedMultiFieldCache( defaultByteArray( (long) 10, INSTANCE ), ID_SLOT_SIZE, 5, 1 );
+        PackedMultiFieldCache cache = new PackedMultiFieldCache( defaultByteArray( 10, INSTANCE ), ID_SLOT_SIZE, 5, 1 );
         int index = 0;
 
         // WHEN
@@ -67,7 +67,7 @@ class PackedMultiFieldCacheTest
     void shouldBeAbleToChangeSlotSize()
     {
         // GIVEN
-        PackedMultiFieldCache cache = new PackedMultiFieldCache( defaultByteArray( (long) 20, INSTANCE ), 5, 1 );
+        PackedMultiFieldCache cache = new PackedMultiFieldCache( defaultByteArray( 20, INSTANCE ), 5, 1 );
         int index = 10;
         assertThrows( IllegalArgumentException.class, () -> cache.put( index, 2, 0 ) );
 
@@ -83,7 +83,7 @@ class PackedMultiFieldCacheTest
     void shouldHandleTwoIdsAndFourBooleans()
     {
         // given
-        PackedMultiFieldCache cache = new PackedMultiFieldCache( defaultByteArray( (long) 10, INSTANCE ), ID_SLOT_SIZE, ID_SLOT_SIZE, 1, 1, 1, 1 );
+        PackedMultiFieldCache cache = new PackedMultiFieldCache( defaultByteArray( 10, INSTANCE ), ID_SLOT_SIZE, ID_SLOT_SIZE, 1, 1, 1, 1 );
         int index = 3;
 
         // when
